src/django/requirements.txt
Django>=2.2.7
djangorestframework>=3.10.3
pandas>=0.25.3
scikit-learn>=0.21.3
mysqlclient>=1.3.2
pickle5
#asgiref==3.4.1
#joblib==0.14.0
#numpy==1.17.3
#pytz==2021.1
#scipy==1.7.1
#six==1.16.0
#sqlparse==0.4.1
#threadpoolctl==2.2.0
#gunicorn==20.1.0